# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PicHub-Aggregator 图片聚合分发服务。管理接口支持会话和 Bearer token 鉴权，图片分发接口公开。",
        "title": "PicHub Aggregator API",
        "termsOfService": "https://github.com/untitled572/pichub-aggregator",
        "contact": {
            "name": "PicHub"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.5"
    },
    "basePath": "/",
    "paths": {
...
        "/api/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "scope 含 images 时导出 ZIP (含 manifest.json + 已保存图片)，否则导出 JSON",
                "produces": [
                    "application/json",
                    "application/zip"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "导出备份",
                "parameters": [
                    {
                        "enum": [
                            "config",
                            "stats",
                            "images"
                        ],
                        "type": "string",
                        "default": "config,stats,images",
                        "description": "导出范围(逗号分隔)",
                        "name": "scope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "备份文件(JSON 或 ZIP)",
                        "schema": {
                            "$ref": "#/definitions/model.ExportManifest"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "导出数据源规则，自动剔除敏感 headers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "导出规则",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "支持 multipart 上传 ZIP/JSON 文件或直接提交 manifest JSON",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "导入备份",
                "parameters": [
                    {
                        "type": "file",
                        "description": "备份文件 (multipart)",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "description": "备份 JSON (非 multipart 时)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.ExportManifest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ExportManifest": {
            "type": "object",
            "properties": {
                "exported_at": {
                    "type": "string"
                },
                "saved_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SavedImage"
                    }
                },
                "scope": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "settings": {
                    "$ref": "#/definitions/model.Settings"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Source"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/model.StatsExportData"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.ImageHistoryRecord": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_saved": {
                    "type": "boolean"
                },
                "source_id": {
                    "type": "integer"
                },
                "source_name": {
                    "type": "string"
                }
            }
        },
        "model.QueryParam": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "model.SavedImage": {
            "type": "object",
            "properties": {
                "file_id": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "format": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "orientation": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "saved_at": {
                    "type": "string"
                },
                "source_name": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "model.Settings": {
            "type": "object",
            "properties": {
                "admin_password": {
                    "type": "string"
                },
                "admin_username": {
                    "type": "string"
                },
                "bound_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cache_max_images": {
                    "type": "integer"
                },
                "cache_max_mb": {
                    "type": "integer"
                },
                "cache_ttl": {
                    "type": "integer"
                },
                "health_check_interval": {
                    "type": "integer"
                },
                "login_enabled": {
                    "type": "boolean"
                },
                "max_history_records": {
                    "type": "integer"
                },
                "min_resolution": {
                    "type": "string"
                },
                "pool_size": {
                    "type": "integer"
                },
                "precache_count": {
                    "type": "integer"
                },
                "proxy_enabled": {
                    "type": "boolean"
                },
                "proxy_mode": {
                    "type": "boolean"
                },
                "proxy_url": {
                    "type": "string"
                },
                "rate_limit": {
                    "type": "integer"
                },
                "rate_limit_window": {
                    "type": "integer"
                },
                "saved_images_dir": {
                    "type": "string"
                },
                "session_hours": {
                    "type": "integer"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "model.Source": {
            "type": "object",
            "properties": {
                "avg_latency": {
                    "type": "integer"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "default_query": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "fail_count": {
                    "type": "integer"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "json_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueryParam"
                    }
                },
                "resp_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "success_rate": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "model.StatsExportData": {
            "type": "object",
            "properties": {
                "image_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ImageHistoryRecord"
                    }
                },
                "stats_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StatsRequestRow"
                    }
                },
                "stats_source": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StatsSourceRow"
                    }
                },
                "stats_tag": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StatsTagRow"
                    }
                }
            }
        },
        "model.StatsRequestRow": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "model.StatsSourceRow": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "hit_count": {
                    "type": "integer"
                },
                "source_id": {
                    "type": "integer"
                },
                "source_name": {
                    "type": "string"
                }
            }
        },
        "model.StatsTagRow": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "tag_id": {
                    "type": "string"
                }
            }
        },
        "model.Tag": {
            "type": "object",
            "properties": {
                "exclusive": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "system": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}